+++ example/src/component/menu1/Menu1.jsx
@@ -1,18 +1,23 @@
 import React from 'react';
-import { NavLink, useLocation } from 'react-router-dom';
+import { useHistory, useLocation } from 'react-router-dom';
 import * as Icon from 'react-bootstrap-icons';
 
 import { ReactDrilldownMenu } from 'react-drilldown-menu';
 
 export default function Menu1() {
   const location = useLocation();
+  const history = useHistory();
+
+  const handleLeafNodeClick = (newPath) => {
+    history.push(newPath);
+  };
+
   return (
     <div className="menu">
       <ReactDrilldownMenu
         name="Menu 1"
         icon={<Icon.FileBinaryFill />} // asdads
         activeLink={location.pathname}
-        navLink={NavLink}
         defaultLeafIcon={<Icon.CircleFill />}
         defaultBranchIcon={<Icon.SquareFill />}
         links={{
@@ -34,7 +39,20 @@ export default function Menu1() {
               },
             },
           },
+          'node-2': {
+            label: 'Node 2',
+            links: {
+              'page-3': {
+                label: 'Page 3',
+              },
+            },
+          },
+
+          'page-6': {
+            label: 'Page 6',
+          },
         }}
+        onLeafNodeClick={handleLeafNodeClick}
       />
     </div>
   );