+++ src/theme.js
@@ -1,5 +1,10 @@
 export default {
   selectionBackgroundColor: '#5098cd',
+  linkIconColor: '#888888',
+  linkTextColor: '#111111',
+  activeLinkTextColor: '#ffffff',
+  activeLinkIconColor: '#ffffff',
+  hoverLinkBackgroundColor: '#f5f5f5',
 
   linkHeightNumber: 37,
   linkHeight: '37px',