+++ History.md
@@ -1,8 +1,15 @@
+# 1.1.3 / 16/Nov/2020
+
+- Fix the Link not effecting page change
+- Update docs
+- Add theme variables
+- Add new handler for page change
+
 # 1.1.2 / 14/Nov/2020
 
 - Add Changelog
 - Fix the Link not effecting page change
-- Remove react-router-dom from lib dependency.
+- Remove react-router-dom from lib dependency
 - Remove styled components
 - Update Readme with Basic usage example (Soon will update information for all the props)
 - Add keywords